why-work/requirements.txt
fastapi
uvicorn
requests
yfinance
jinja2
python-multipart
aiohttp
tradingview-ta
gunicorn
sqlalchemy
aiomysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
greenlet
